testing out a new approach utilizing canvaslayers
Rather than using a viewportcontainer and a subviewport for each tilemap, I am just creating a canvaslayer for each "layer" of the level.

Under this canvaslayer I have a subviewport and a tilemap.
I set the custom viewport of the canvaslayer to its subviewport in a script.

I then display the contents of each layer with a texturerect